Work for ref #100 where :default attribute values can take procs (but parent arg is nil)
This initial support will currently not properly pass the parent object into the first 
argument of the proc. It will always be a FakeClass that will complain (warn) for any
access to any method within it. 
This partial feature is only going to exist until we can retfactor the attribute and type class loads so that they can pass that information along (requires sizable refactoring)

Signed-off-by: Josep M. Blanquer <blanquer@rightscale.com>