Add present_if option
When dumping a model, each attribute will be replaced with nil if it has
a present_if condition and that condition is not met. The allowed
conditions are the same as for required_if:

* A string key, which passes if that key's value is not nil.
* A one-item hash, whose key is the path in the object, and whose value
  is any ===-compatible object (string, regex, proc, etc.).
* A proc, which is passed the object.

Like required_if, this cannot be used in conjunction with the required
option. It can be used with the required_if option, but it should only
be so in a way that they are compatible (for instance, if both have the
same condition, then the attribute will only be present when it is
required).

Signed-off-by: Sean McGivern <sean.mcgivern@rightscale.com>